+++ src/main/java/com/hongikbros/jobmanager/common/domain/BaseEntity.java
@@ -1,45 +1,27 @@
 package com.hongikbros.jobmanager.common.domain;
 
-import java.io.Serializable;
-import java.time.LocalDateTime;
-
 import javax.persistence.EntityListeners;
 import javax.persistence.MappedSuperclass;
 
 import org.springframework.data.annotation.CreatedBy;
-import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedBy;
-import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
 @MappedSuperclass
 @EntityListeners(AuditingEntityListener.class)
-public class BaseEntity implements Serializable {
+public class BaseEntity extends BaseTimeEntity {
+
     @CreatedBy
     private String createdBy;
 
-    @CreatedDate
-    private LocalDateTime createdDate;
-
     @LastModifiedBy
     private String lastModifiedBy;
-
-    @LastModifiedDate
-    private LocalDateTime lastModifiedDate;
 
     public String getCreatedBy() {
         return createdBy;
     }
 
-    public LocalDateTime getCreatedDate() {
-        return createdDate;
-    }
-
     public String getLastModifiedBy() {
         return lastModifiedBy;
-    }
-
-    public LocalDateTime getLastModifiedDate() {
-        return lastModifiedDate;
     }
 }
